Updating getSome
This patch introduces some optimizations for the getSome object. The
total number of images pulled from instagram during the ajax call
are cached in the object and later referenced in the new method
called next(). This new method handles calls to impress.next() and
refreshes the browser to preform the next load of Instagram images.